Update patch set 1
Abandoned

i think that was a manifest bug. we will have to build the kernel with this in future, but not right now. if we can't find & fix GCC before then, it's easy to resurrect this. in the meantime, let's stick with this version in case there are other problems left to find...

Patch-set: 1
Status: abandoned