Update patch set 1
Patch Set 1:

This is rebuilding prebuilts after reserving x18. Dan, isn't there an option in the script for setting a bug number? If not, maybe we should just manually add the bug number to these patches, so it is understandable for future versions of ourselves. ;)

Patch-set: 1
Reviewer: Gerrit User 1010056 <1010056@85c56323-6fa9-3386-8a01-6480fb634889>